move is_valid_blobhash to lbrynet.blob.blob_file
