+++ lbrynet/p2p/client/StandaloneBlobDownloader.py
@@ -4,7 +4,8 @@ from lbrynet.p2p.client.BlobRequester import BlobRequester
 from lbrynet.p2p.client.ConnectionManager import ConnectionManager
 from lbrynet.p2p.client.DownloadManager import DownloadManager
 from lbrynet.p2p.Error import InvalidBlobHashError, DownloadSDTimeout
-from lbrynet.utils import is_valid_blobhash, safe_start_looping_call, safe_stop_looping_call
+from lbrynet.blob.blob_file import is_valid_blobhash
+from lbrynet.utils import safe_start_looping_call, safe_stop_looping_call
 from twisted.python.failure import Failure
 from twisted.internet import defer
 from twisted.internet.task import LoopingCall
